Написал тест на отображение пустой страницы
